pyproject.toml
[project]
name = "scheduled-task-reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.1",
    "aiomysql>=0.2.0",
    "aiosmtplib>=4.0.2",
    "apscheduler>=3.11.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.2",
    "fastapi[all]>=0.118.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.37.0",
]

[tool.aerich]
tortoise_orm = "app.models.mysql_config.MYSQL_TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
